+++ commands/transfer_funds_test.go
@@ -43,6 +43,10 @@ func TestTransferFunds(t *testing.T) {
 	capNegative := "-1"
 	capZero := "0"
 
+	tooLongTransferInterval := int32(101)
+	zeroTransferInterval := int32(0)
+	negativeTransferInterval := int32(-1)
+
 	cases := []struct {
 		transfer  commandspb.Transfer
 		errString string
@@ -1199,7 +1203,7 @@ func TestTransferFunds(t *testing.T) {
 				Amount:    "1",
 				Reference: "testing",
 			},
-			errString: "transfer.kind.dispatch_strategy.window_length (must bet between 1 and 100)",
+			errString: "transfer.kind.dispatch_strategy.window_length (must be between 1 and 100)",
 		},
 		{
 			transfer: commandspb.Transfer{
@@ -1298,6 +1302,84 @@ func TestTransferFunds(t *testing.T) {
 				Amount:    "1",
 				Reference: "testing",
 			},
+		},
+		{
+			transfer: commandspb.Transfer{
+				FromAccountType: vega.AccountType_ACCOUNT_TYPE_GENERAL,
+				ToAccountType:   vega.AccountType_ACCOUNT_TYPE_REWARD_AVERAGE_POSITION,
+				Kind: &commandspb.Transfer_Recurring{
+					Recurring: &commandspb.RecurringTransfer{
+						StartEpoch: 10,
+						EndEpoch:   ptr.From(uint64(11)),
+						Factor:     "1",
+						DispatchStrategy: &vega.DispatchStrategy{
+							AssetForMetric:   "",
+							Metric:           vega.DispatchMetric_DISPATCH_METRIC_AVERAGE_POSITION,
+							EntityScope:      vega.EntityScope_ENTITY_SCOPE_INDIVIDUALS,
+							IndividualScope:  vega.IndividualScope_INDIVIDUAL_SCOPE_IN_TEAM,
+							WindowLength:     100,
+							TransferInterval: &tooLongTransferInterval,
+						},
+					},
+				},
+				To:        "84e2b15102a8d6c1c6b4bdf40af8a0dc21b040eaaa1c94cd10d17604b75fdc35",
+				Asset:     "080538b7cc2249de568cb4272a17f4d5e0b0a69a1a240acbf5119d816178daff",
+				Amount:    "1",
+				Reference: "testing",
+			},
+			errString: "transfer.kind.dispatch_strategy.transfer_interval (must be between 1 and 100)",
+		},
+		{
+			transfer: commandspb.Transfer{
+				FromAccountType: vega.AccountType_ACCOUNT_TYPE_GENERAL,
+				ToAccountType:   vega.AccountType_ACCOUNT_TYPE_REWARD_AVERAGE_POSITION,
+				Kind: &commandspb.Transfer_Recurring{
+					Recurring: &commandspb.RecurringTransfer{
+						StartEpoch: 10,
+						EndEpoch:   ptr.From(uint64(11)),
+						Factor:     "1",
+						DispatchStrategy: &vega.DispatchStrategy{
+							AssetForMetric:   "",
+							Metric:           vega.DispatchMetric_DISPATCH_METRIC_AVERAGE_POSITION,
+							EntityScope:      vega.EntityScope_ENTITY_SCOPE_INDIVIDUALS,
+							IndividualScope:  vega.IndividualScope_INDIVIDUAL_SCOPE_IN_TEAM,
+							WindowLength:     100,
+							TransferInterval: &zeroTransferInterval,
+						},
+					},
+				},
+				To:        "84e2b15102a8d6c1c6b4bdf40af8a0dc21b040eaaa1c94cd10d17604b75fdc35",
+				Asset:     "080538b7cc2249de568cb4272a17f4d5e0b0a69a1a240acbf5119d816178daff",
+				Amount:    "1",
+				Reference: "testing",
+			},
+			errString: "transfer.kind.dispatch_strategy.transfer_interval (must be between 1 and 100)",
+		},
+		{
+			transfer: commandspb.Transfer{
+				FromAccountType: vega.AccountType_ACCOUNT_TYPE_GENERAL,
+				ToAccountType:   vega.AccountType_ACCOUNT_TYPE_REWARD_AVERAGE_POSITION,
+				Kind: &commandspb.Transfer_Recurring{
+					Recurring: &commandspb.RecurringTransfer{
+						StartEpoch: 10,
+						EndEpoch:   ptr.From(uint64(11)),
+						Factor:     "1",
+						DispatchStrategy: &vega.DispatchStrategy{
+							AssetForMetric:   "",
+							Metric:           vega.DispatchMetric_DISPATCH_METRIC_AVERAGE_POSITION,
+							EntityScope:      vega.EntityScope_ENTITY_SCOPE_INDIVIDUALS,
+							IndividualScope:  vega.IndividualScope_INDIVIDUAL_SCOPE_IN_TEAM,
+							WindowLength:     100,
+							TransferInterval: &negativeTransferInterval,
+						},
+					},
+				},
+				To:        "84e2b15102a8d6c1c6b4bdf40af8a0dc21b040eaaa1c94cd10d17604b75fdc35",
+				Asset:     "080538b7cc2249de568cb4272a17f4d5e0b0a69a1a240acbf5119d816178daff",
+				Amount:    "1",
+				Reference: "testing",
+			},
+			errString: "transfer.kind.dispatch_strategy.transfer_interval (must be between 1 and 100)",
 		},
 	}